+++ src/main/java/mech/mania/engine/game/items/StatusModifier.java
@@ -2,15 +2,15 @@ package mech.mania.engine.game.items;
 
 public class StatusModifier {
     private int speedChange;
-    private double healthChange;
-    private double experienceChange;
-    private double magicDamageChange;
-    private double physicalDamageChange;
-    private double magicDefenseChange;
-    private double physicalDefenseChange;
+    private int healthChange;
+    private int experienceChange;
+    private int magicDamageChange;
+    private int physicalDamageChange;
+    private int magicDefenseChange;
+    private int physicalDefenseChange;
 
-    public StatusModifier(int speedChange, double healthChange, double experienceChange, double magicDamageChange,
-                          double physicalDamageChange, double magicDefenseChange, double physicalDefenseChange) {
+    public StatusModifier(int speedChange, int healthChange, int experienceChange, int magicDamageChange,
+                          int physicalDamageChange, int magicDefenseChange, int physicalDefenseChange) {
         this.speedChange = speedChange;
         this.healthChange = healthChange;
         this.experienceChange = experienceChange;
@@ -28,51 +28,51 @@ public class StatusModifier {
         this.speedChange = speedChange;
     }
 
-    public double getHealthChange() {
+    public int getHealthChange() {
         return healthChange;
     }
 
-    public void setHealthChange(double healthChange) {
+    public void setHealthChange(int healthChange) {
         this.healthChange = healthChange;
     }
 
-    public double getExperienceChange() {
+    public int getExperienceChange() {
         return experienceChange;
     }
 
-    public void setExperienceChange(double experienceChange) {
+    public void setExperienceChange(int experienceChange) {
         this.experienceChange = experienceChange;
     }
 
-    public double getMagicDamageChange() {
+    public int getMagicDamageChange() {
         return magicDamageChange;
     }
 
-    public void setMagicDamageChange(double magicDamageChange) {
+    public void setMagicDamageChange(int magicDamageChange) {
         this.magicDamageChange = magicDamageChange;
     }
 
-    public double getPhysicalDamageChange() {
+    public int getPhysicalDamageChange() {
         return physicalDamageChange;
     }
 
-    public void setPhysicalDamageChange(double physicalDamageChange) {
+    public void setPhysicalDamageChange(int physicalDamageChange) {
         this.physicalDamageChange = physicalDamageChange;
     }
 
-    public double getMagicDefenseChange() {
+    public int getMagicDefenseChange() {
         return magicDefenseChange;
     }
 
-    public void setMagicDefenseChange(double magicDefenseChange) {
+    public void setMagicDefenseChange(int magicDefenseChange) {
         this.magicDefenseChange = magicDefenseChange;
     }
 
-    public double getPhysicalDefenseChange() {
+    public int getPhysicalDefenseChange() {
         return physicalDefenseChange;
     }
 
-    public void setPhysicalDefenseChange(double physicalDefenseChange) {
+    public void setPhysicalDefenseChange(int physicalDefenseChange) {
         this.physicalDefenseChange = physicalDefenseChange;
     }
 }
